Add comment feature

+++ app/Http/Controllers/CommentController.php
@@ -7,18 +7,30 @@ use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create()
     {
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
         return back();
     }
     public function delete($id)
     {
+
         $comment = Comment::find($id);
-        $comment->delete();
-        return back();
+        if ($comment->user_id == auth()->user()->id) {
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error', 'Unauthorized');
+        }
     }
 }
